vae-requirements.txt
absl-py==2.3.1
academictorrents==2.3.3
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
array_record==0.7.2
arrow==1.3.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.3.0
av==15.0.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.13.4
bencode.py==2.0.0
bitstring==3.1.5
bleach==6.2.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.2
cycler==0.12.1
dataclasses==0.6
debugpy @ file:///croot/debugpy_1736267418885/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
decord==0.6.0
defusedxml==0.7.1
diffusers==0.34.0
dm-tree==0.1.9
docopt==0.6.2
docstring_parser==0.17.0
einops==0.8.1
etils==1.13.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
fastjsonschema==2.21.1
filelock==3.18.0
fire==0.7.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
future==0.16.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
gql==3.5.3
graphql-core==3.2.6
halo==0.0.31
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
imageio==2.37.0
immutabledict==4.2.1
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources==6.5.2
iniconfig==2.1.0
inquirerpy==0.3.4
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython==8.12.3
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1748333051527/work
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.14.3
log-symbols==0.0.14
lpips==0.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mistune==3.1.3
modelscope==1.28.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
# Editable install with no version control (myutils==0.1)
-e /cs/student/projects1/2023/rcheng/utils
narwhals==1.47.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==1.4.4
pandocfilters==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pfzy==0.3.4
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.3.0
pipreqs==0.5.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
pluggy==1.6.0
promise==2.3
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
propcache==0.3.2
protobuf==6.31.1
psutil @ file:///croot/psutil_1736367091698/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pudb==2025.1
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing==3.2.3
PyPubSub==3.3.0
pytest==8.4.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytorch-lightning==2.5.2
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
safetensors==0.5.3
scikit-image==0.21.0
scipy==1.16.0
shellingham==1.5.4
simple-parsing==0.1.7
simsimd==6.5.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spinners==0.0.24
SSIM-PIL==1.0.14
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
streamlit==1.46.1
stringzilla==3.12.5
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorflow-datasets==4.9.9
tensorflow-metadata==1.17.2
termcolor==3.1.0
tifffile==2025.6.11
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchmetrics==1.7.4
torchvision==0.22.1
tornado @ file:///croot/tornado_1748956929273/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transform==1.2.0
transform_tools==1.0.5
transformers==4.53.2
triton==3.3.1
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
urwid==3.0.2
urwid_readline==0.15.1
validators==0.35.0
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
webencodings==0.5.1
wrapt==1.17.2
xxhash==3.5.0
yarg==0.1.9
yarl==1.20.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
